removing unnecesary /node_modules/ path and renaming Player files
