+++ src/components/Player/index.js
@@ -1,0 +1,20 @@
+import React from "react"
+import styles from "./Player.module.css"
+import PropTypes from "prop-types"
+
+function Player({name, removePlayer}) {
+  if (!name) return null
+  return (
+    <div className={styles.player}>
+      <span>{name}</span>
+      <button onClick={() => removePlayer(name)}>X</button>
+    </div>
+  )
+}
+
+Player.propTypes = {
+  name: PropTypes.string,
+  removePlayer: PropTypes.func,
+}
+
+export default Player
