+++ src/components/Player/index.stories.js
